// client/src/app/shared/models/ClientLocation.ts
export interface IClientLocation {
  id: number;
  companyName: string;
  managerName: string;
  contactName: string;
  address1: string;
  address2: string;
  address3: string;
  address4: string;
  address5: string;
  contactNumber: string;
  agencyId: number;
}
export class ClientLocation {
  id: number;
  companyName: string;
  managerName: string;
  contactName: string;
  address1: string;
  address2: string;
  address3: string;
  address4: string;
  address5: string;
  contactNumber: string;
  agencyId: number;
}