// client/src/app/shared/models/JobConformed.ts
export interface JobConformed {
  id: number;
  gradeDetails: string;
  photoUrl: string;
  message: string;
  jobDate: string;
  startTime: string;
  endTime: string;
  jobAddress: string;
  paymentDescription: string;
  comment: string;
  rating: number;
  finishShift: boolean;
  lostShift: boolean;
  timeDetail: string;
  timeDetailId: number;
  paymentType: string;
  paymentTypeId: number;
  candidate: string;
  candidateId: number;
  agency: string;
  agencyId: number;
  jobToRequest: string;
  jobToRequestId: number;
  gradeId: number;
  grade: string;
  clientLocation: string;
  clientLocationId: number;
  shiftState: string;
  shiftStateId: number;
  appUserPosted: string;
  appUserPostedId: string;
  appUserCandidate: string;
  appUserCandidateId: string;
  ariaId: number;
  aria: string;
  attributeDetail: string;
  attributeDetailId: number;
  agencyPhotoUrl: string;
  jobType: string;
  jobTypeId: number;
}