// client/src/app/shared/models/candidate-jobs-invite.ts
export interface ICandidateJobInvite {
  id: number;
  isSelected: boolean;
  message: string;
  accept?: boolean;
  reject?: boolean;
  publishDate: string;
  publishTime: string;
  responseDate?: Date;
  respondTime?: Date;
  jobDate: string;
  startTime: string;
  endTime: string;
  jobAddress: string;
  paymentType: string;
  paymentTypeId: number;
  candidate: string;
  candidateId: number;
  agency: string;
  agencyId: number;
  avatarAgencyUrl: string;
  jobToRequestId: number;
  gradeId: number;
  grade: string;
  shiftState: string;
  shiftStateId: number;
  ariaId: number;
  aria: string;
  attributeDetail: string;
  attributeDetailId: number;
  jobType: string;
  jobTypeId: number;
}
export class CandidateJobInvite implements ICandidateJobInvite{
  id: number;
  isSelected: boolean;
  message: string;
  accept?: boolean;
  reject?: boolean;
  publishDate: string;
  publishTime: string;
  responseDate?: Date;
  respondTime?: Date;
  jobDate: string;
  startTime: string;
  endTime: string;
  jobAddress: string;
  paymentType: string;
  paymentTypeId: number;
  candidate: string;
  candidateId: number;
  agency: string;
  agencyId: number;
  avatarAgencyUrl: string;
  jobToRequestId: number;
  gradeId: number;
  grade: string;
  shiftState: string;
  shiftStateId: number;
  ariaId: number;
  aria: string;
  attributeDetail: string;
  attributeDetailId: number;
  jobType: string;
  jobTypeId: number;
}