// client/src/app/shared/models/CandidateDocument.ts
export interface ICandidateDocument {
    url: string;
    description: string;
    dateAdded: string;
    publicId: string;
    candidateId: number;
    id: number;
}
export class CandidateDocument {
    url: string;
    description: string;
    dateAdded: string;
    publicId: string;
    candidateId: number;
}